Refact: Splitted AndroidNfcReaderImpl.kt into AndroidNfcReaderPostNImpl.kt and AndroidNfcReaderPreNImpl.kt in order to take advantage of initstate service refactoring.
